requirements.txt
google-generativeai>=0.8.3
python-dotenv==1.0.0
PyPDF2==3.0.1
llama-cpp-python>=0.3.0
httpx>=0.25.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
